Pretending to be an Adafruit board to get it working with 4mb.

+++ include/EnvironmentDisplay.h
@@ -6,7 +6,7 @@ namespace EnvironmentDisplay {
     void displayTime(const struct tm *clockTime);
     void displayTemperature(float temperature);
     void displayHumidity(long humidity);
-    void displayBrightness(float light);
+    void displayBrightness(long light);
     void displayGpsStatus(bool gpsStatus);
     void displayMessage(const char *message);
     void clearMessage(void);
